Fix input format
When the input is a flat Vector{F}, it is assumed to be a
constant-in-time input state. That is, it is assume that the array
represents the state of the inputs, and the same values are used at all
time steps.

What we really want is a Vector{Vector{F}}, meaning an array of arrays.
Each of the elements of the input should be an array represent the
temperature (a length 1 vector). There should be one such vector for
each timestep.