Handle 10W_SD differently
To improve performance, we will do two things. First, we will ensure
that given any random initial state, the system will fall into
more-or-less the all-zero state after 10 weeks of warm. That contributes
to the loss. Second, we will assume that all of the cold and cold/warm
simulations start in the all zero state. We will not simulate the
preceeding 10 weeks of warm. This cuts the runtime in half, and doesn't
really hurt the fit.